盤面の黒線のレイアウトを修正
